handle_downlink: allow map returned by downlink function
Up to now we could only set the 'data' field via the binary return value of
the downlink function.

With the patch we allow the downlink function to return either a binary or
a map.

If a binary is returned, it is used to update the 'data' value.

If a map is returned, this map replaces the original map passed to the
downlink function. All relevant download fields like devaddr, deveui, port,
pending, receipt, and data can be set.

Here is a simple example:

    fun(Vars) ->
        Vars#{
            deveui => <<16#3065ECFEFF6FC458:64>>,
            port => 43,
            data => <<16#02, 16#0113:16/little>>
        }
    end.

Closes #572.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>