MQTT connector: retain flag
The retain flag in the uplink message to an MQTT server signals that the
message content shall be kept and presented to any client subscribing to
the topic at a later moment.

With the patch the retain flag can be set in the uplink function of the
handler.

retain = true       - create a retained record
retain = delete     - delete the retained record and create a non-retained
                      record.
none or other value - create a non-retained record

To avoid sending irrelevant data the field 'retain' is removed from the
message sent to the MQTT server.

Closes #579

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>