pass TERM signal to underlying process and improve shutdown
The lorawan-server script doesn't forward TERM signal to underlying process. 

Normally the underlying process must be started from the starting script with "exec", but only if it is a binary executable (which "erl" is not; from tests carried out it works, but two "zombie" processes remain active). In this modification, a trap is activated that forwards the term signal.

In solution to #843 issue